Find the common ancestor in a merge-free history
This commit begins work on a `merge` command for unifying two divergent
branches back together. The first step in this process is to find the
common ancestor of the two branch tips. For example, in this history,
where `parent` links point from right to left, the `master` branch
points to commit D and `topic` to commit G:

    A   B   C   D
    o---o---o---o [master]
         \
          o---o---o [topic]
          E   F   G

If we have `master` checked out and run `merge topic`, we need to locate
B as the best common ancestor of D and G -- the last commit that each
history diverged from.

In a history containing no merges, the commits form a tree. There is a
unique path from a commit to any of its ancestors, and there is a unique
best common ancestor of any two commits. To find it, it's sufficient to
walk backwards from the two commits and return the first commit that's
reachable from both of them.

`Merge::CommonAncestors` does this by flagging the HEAD commit as
`parent1` and the merge target as `parent2`, and places both commits in
a date-ordered queue as in `RevList`. It processes each item from the
queue by propagating its flags to its parents. As soon as a commit with
both flags is found, we return it.