Don't log a commit when it is treesame to one of its parents
The command `log <path>` results in a list of commits that change the
given path. This currently works by computing the tree diff of the
commit against its (first) parent, pruned to the path of interest. If
the tree diff is empty, then that file is unchanged by the commit, and
the commit is "treesame" and excluded from the results.

Generalising this to multiple parents means a commit should not be
printed if it's treesame to *any* of its parents. If this is the case,
it means the merge took its version of the file from one of its parents,
rather than introducing any changes itself. The true change will be in
the commits leading up to the merge.