Write files to the workspace
Handling a merge conflict will sometimes require writing an untracked
file to the workspace, for example in the case of a file-directory
conflict. In this case, the usual migration machinery (e.g. the
`Workspace#apply_migration` method) used for checkouts and merges won't
create the required file, and we need a method for creating ad-hoc files
in the workspace.