Print warnings about merge conflicts
While the `Merge::Resolve` class is detecting conflicts, it should
generate warnings and the `Command::Merge` class should print them, so
the user can see what failed. The logic here is fairly boring
bookkeeping that detects various different cases in the same way that
Git does, and prints one of the following warnings.

When we attempt to merge two conflicting edits:

    Auto-merging <path>

When adding one path in a file-directory conflict:

    Adding <path>

When both sides have added or edited the same file:

    CONFLICT (content|add/add): Merge conflict in <path>

When one side has modified a file that the other deleted:

    CONFLICT (modify/delete): <path> deleted in <v1> and modified in
    <v2>. Version <v2> of <path> left in tree.

When there is a file-directory conflict:

    CONFLICT (file/directory): There is a directory with name <path> in
    <v1>. Adding <path> as <path>~<v2>.