Don't remove files with unstaged changes
The `rm` command tries to be safer than the normal rm(1) program, in
that it tries to prevent you losing data. The easiest way to irrevocably
lose data is by deleting things from the workspace that have not been
saved to the database, by being added to the index.

Therefore, we use the `Repository::Inspector` class to check whether the
file on disk matches the index entry. If not, there are unstaged changes
and we refuse to remove the file. This check is only done if the file
exists -- if it's been deleted already, there is no data that would be
lost by running the `rm` operation.

All arguments are checked for unstaged changes first, and they are only
removed if it's safe to remove all of them.