Reset index entries to their version in HEAD
Here we begin the `reset` command, which allows versions of files from
arbitrary commits to be loaded into the index. When run with the names
of files as input, it replaces the current index entry for each file
with its version from HEAD, effectively "unstaging" the file. The copy
in the workspace remains unmodified, and HEAD does not move.

The index entries restored from HEAD do not carry `stat(2)` data,
because they do not necessarily reflect the file in the workspace.