Add the `--hard` mode to the `reset` command
When run with a mode rather than a list of pathnames, the behaviour of
`reset` is:

- `soft`: move HEAD so that it points at the given commit

- `mixed`: as `soft`, plus restore the index to contain the contents of
  the given commit

- `hard`: as `mixed`, plus restore the workspace to match the index and
  HEAD as of the given commit

However, implementing `hard` mode is not as simple as just running the
`mixed` code paths and then doing some more work to update the
workspace. This is for a couple of reasons.

First, updating the index may cause some files in the workspace to
become untracked. These files should be deleted, but files that were
untracked before the reset should remain. So, whereas the easiest way to
update the index is to just remove the previous entries under a path and
replace them, this approach won't work for the workspace. Either you
clear the workspace and completely rebuild it, which would delete
untracked files, or you compare it to the index and fix any differences,
but as we said, this would result in some deleted files persisting as
untracked files.

Second, there are efficiency reasons to prefer updating the workspace
before doing anything else. If we compare it to the index before the
index is updated, we can work out just those paths that differ and amend
them. Plus, since we're making the workspace match the index, it would
be good if the index entries contained `stat(2)` data for the newly
updated files.

A better way to look at hard reset is as a distinct operation from the
other modes, which eliminates any differences that the index and
workspace have from the given commit. We can amend the
`Repository::Status` class to take a commit ID as input and use this in
place of HEAD as the base for its comparisons. Any changes reported are
then removed from the workspace and index, using the new filestat data
to populate the index.

Adding this operation as a method to `Repository` means we can invoke it
from other commands, which will be useful for implementing `--abort` on
various merge operations.